+++ ckan/views/dataset.py
@@ -27,7 +27,7 @@ from ckan.common import _, config, g, request
 from ckan.views.home import CACHE_PARAMETERS
 from ckan.lib.plugins import lookup_package_plugin
 from ckan.lib.search import (
-    SearchError, SearchQueryError, SearchIndexError, SolrConnectionError
+    SearchError, SearchQueryError, SearchIndexError, SearchConnectionError
 )
 from ckan.types import Context, Response
 
@@ -337,13 +337,13 @@ def search(package_type: str) -> str:
             _(u'Invalid search query: {error_message}')
             .format(error_message=str(se))
         )
-    except (SearchError, SolrConnectionError) as se:
-        if isinstance(se, SolrConnectionError):
+    except (SearchError, SearchConnectionError) as se:
+        if isinstance(se, SearchConnectionError):
             base.abort(500, se.args[0])
 
         # May be bad input from the user, but may also be more serious like
-        # bad code causing a SOLR syntax error, or a problem connecting to
-        # SOLR
+        # bad code causing a query syntax error, or a problem connecting to
+        # the search backend
         log.error(u'Dataset search error: %r', se.args)
         extra_vars[u'query_error'] = True
         extra_vars[u'search_facets'] = {}